fix: substitui double por decimal no preço dos produtos

+++ Data/IProdutoRepository.cs
@@ -7,6 +7,6 @@ public interface IProdutoRepository
     void EnsureTabelaProdutos();
     void Criar(Produto produto);
     List<Produto> ListarProdutos();
-    bool Atualizar(string nome, double preco);
+    bool Atualizar(string nome, decimal preco);
     bool Deletar(string nome);
 }
